Update README.md
I think that it is confusing, to call the `main` module `base` module. For instance: in my case I had a complex module-tree, and the "base" in my understanding it the place where I want to use the library. Doing so without importing the lib in the main module will produce an error, see https://github.com/farengeyt451/ngx-tippy-wrapper/issues/31